Refactor apply styles and expose all utils
