+++ src/index.ts
@@ -3,3 +3,4 @@ export * from './internal/index.ts';
 export * from './hooks/index.ts';
 export * from './components/index.ts';
 export * from './context/index.ts';
+export * from './utils/index.ts';
