+++ src/utils/index.ts
@@ -1,0 +1,7 @@
+export * from './apply-fill-and-stroke-style.ts';
+export * from './get-recommended-pixel-ratio.ts';
+export * from './is-own-property-of.ts';
+export * from './is-valid-stroke-cap.ts';
+export * from './is-valid-stroke-join.ts';
+export * from './resolve-styles.ts';
+export * from './type-guards.ts';
